+++ docs/changelog.rst
@@ -41,6 +41,7 @@ Fixes:
   to Bombardment.
 * :doc:`/plugins/play`: Add a ``relative_to`` config option. Thanks to
   BrainDamage.
+* Fix a crash when a MusicBrainz release has zero tracks.
 
 .. _discogs_client: https://github.com/discogs/discogs_client